@@ -11,19 +11,19 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@aws-cdk/assert": "^1.17.1",
-    "@aws-cdk/aws-ssm": "^1.17.1",
-    "@types/jest": "^24.0.23",
-    "@types/node": "12.12.11",
-    "aws-cdk": "^1.17.1",
-    "jest": "^24.9.0",
-    "ts-jest": "^24.1.0",
-    "ts-node": "^8.5.2",
+    "@aws-cdk/assert": "^1.119.0",
+    "@aws-cdk/aws-ssm": "^1.119.0",
+    "@types/jest": "^27.0.1",
+    "@types/node": "16.7.1",
+    "aws-cdk": "^1.119.0",
+    "jest": "^27.0.6",
+    "ts-jest": "^27.0.5",
+    "ts-node": "^10.2.1",
     "cdk-git-config": "file:../",
-    "typescript": "~3.7.2"
+    "typescript": "~4.3.5"
   },
   "dependencies": {
-    "@aws-cdk/core": "^1.17.1",
-    "source-map-support": "^0.5.16"
+    "@aws-cdk/core": "^1.119.0",
+    "source-map-support": "^0.5.19"
   }
 }
