@@ -1,9 +1,14 @@
 import cdk = require('@aws-cdk/core');
 
 import configParser = require('../../lib/config/configuration')
-import gitToSsm = require('../../lib/gitToSsm')
-import path = require('path')
+import { splitGroupsAtPathDepth }  from '../../lib/config/config-splitter';
 
+import gitToSsm = require('../../lib/constructs/gitToSsm')
+import path = require('path')
+import { ConfigGroup } from '../../lib/config/types/config-types';
+import * as gitConfig from '../../lib';
+
+const ssmRootPath = '/gitconfigstore/root';
 export class ConfigExampleStack extends cdk.Stack {
   constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
@@ -14,16 +19,22 @@ export class ConfigExampleStack extends cdk.Stack {
 
     const configLoader = new configParser.ConfigLoader({
       rootDir: configRoot,
-      ssmRootPath: '/gitconfigstore/root',
-      // filterByCurrentAccount: false, //uncomment to see behavioural differences
+      ssmRootPath: ssmRootPath,
+      filterByCurrentAccount: false, //uncomment to see behavioural differences
     })
 
-    // TODO: cleanup this implementation
-    configLoader.load()
-    configLoader.printConfiguration()
+    const configGroup: ConfigGroup[] = configLoader.load()
+    gitConfig.printConfiguration(configGroup);
+ 
+    const splitGroups = splitGroupsAtPathDepth(configGroup, 1);
+    console.log(splitGroups)
 
-    new gitToSsm.GitToSsm(this, 'SampleConfig', {
-      configuration: configLoader
+    splitGroups.forEach(group => {
+      // To have a stack per group - instantiate that here.
+      new gitToSsm.GitToSsm(this, `SampleConfig${group.filteredByPath}`, {
+        configuration: group.groups,
+        ssmRootPath, 
+      })
     })
   }
 }
