@@ -1,10 +1,11 @@
 import cdk = require('@aws-cdk/core');
 import ssm = require('@aws-cdk/aws-ssm')
 
-import configParser = require('./config/configuration')
+import { ConfigGroup } from '../config/types/config-types';
 
 export interface GitToSsmProps {
-  configuration: configParser.ConfigLoader
+  configuration: ConfigGroup[];
+  ssmRootPath: string;
 }
 
 export class GitToSsm extends cdk.Construct {
@@ -12,15 +13,13 @@ export class GitToSsm extends cdk.Construct {
   constructor(scope: cdk.Construct, id: string, props: GitToSsmProps) {
     super(scope, id);
 
-    const allConfigs = props.configuration.load()
-
-    allConfigs.forEach(configGroup => {
+    props.configuration.forEach(configGroup => {
       
       Object.keys(configGroup.configSets).forEach((key)=> {
 
         const value = configGroup.configSets[key]
 
-        new ssm.StringParameter(this, `${props.configuration.ssmRootPath}-${configGroup.configGroupName}-${key}`, {
+        new ssm.StringParameter(this, `${props.ssmRootPath}-${configGroup.configGroupName}-${key}`, {
           description: `from file: ${configGroup.relativePath}`,
           parameterName: key,
           type: ssm.ParameterType.STRING,
