@@ -1,2 +1,3 @@
-export * from './gitToSsm'
+export * from './constructs/gitToSsm'
 export * from './config'
+export * from './config/config-splitter'
